change name of myart file to index.js and delete console.logs
