+++ pages/index.js
@@ -1,5 +1,17 @@
+import { styled } from "styled-components";
+import PictureList from "../components/PictureList";
+import pictures from "../db.js";
 
-
-export default function HomePage() {
-  return <div></div>;
+export default function MyArt() {
+  return (
+    <>
+      <Heading>My Art</Heading>
+      <PictureList pictures={pictures} />
+    </>
+  );
 }
+
+const Heading = styled.h1`
+  text-align: center;
+  font-size: 2rem;
+`;
