--- pages/myart.js
@@ -1,18 +1,0 @@
-import { styled } from "styled-components";
-import PictureList from "../components/PictureList";
-import pictures from "../db.js";
-
-export default function MyArt() {
-  console.log("test");
-  return (
-    <>
-      <Heading>My Art</Heading>
-      <PictureList pictures={pictures} />
-    </>
-  );
-}
-
-const Heading = styled.h1`
-  text-align: center;
-  font-size: 2rem;
-`;
